fix: add checking for existance of arguments

+++ src/utils/rm.js
@@ -1,8 +1,12 @@
 import fs from 'fs/promises';
 import path from 'path';
 
-const rm = async (filePath) => {
-    const fileToRemove = path.resolve(filePath);
+const rm = async (dir, filePath) => {
+    if (!filePath) {
+        console.log('Invalid input');
+        return
+    }
+    const fileToRemove = path.resolve(dir, filePath);
     try {
         await fs.rm(fileToRemove);
     } catch (err) {
